move input component to shared folder
